add logic rating and finished and read and want

+++ lib/views/widgets/stats_search/read_finish_want_card.dart
@@ -2,10 +2,10 @@ import 'package:book_trail/views/widgets/stats_search/read_finish_want_stats.dar
 import 'package:flutter/material.dart';
 
 class ReadFinishWantCard extends StatelessWidget {
-  final int finishedCount;
-  final int readingCount;
-  final int toReadCount;
-  final int avgRate;
+  final double finishedCount;
+  final double readingCount;
+  final double toReadCount;
+  final double avgRate;
   const ReadFinishWantCard({super.key, required this.finishedCount, required this.readingCount, required this.toReadCount, required this.avgRate});
 
   @override
